LayoutTests/fast/css/lang-pseudo-disconnected-expected.txt

:lang pseudo class should work in a disconnected subtree

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS target.matches(":lang(zh)") is true
PASS target.parentNode.remove(); target.matches(":lang(zh)") is true
PASS target.matches(":lang(fr)") is true
PASS document.body.append(target.parentNode); target.matches(":lang(fr)") is true
PASS successfullyParsed is true

TEST COMPLETE
